pricesentinel/pyproject.toml
[project]
name = "pricesentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "curl-cffi>=0.14.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "pandas>=2.3.3",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "scrapy>=2.14.2",
    "scrapy-playwright>=0.0.46",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
